+++ lib/layrz_models.dart
@@ -6,6 +6,7 @@ import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:layrz_icons/layrz_icons.dart';
+import 'package:layrz_models/src/operations/operations.dart';
 
 // Internal imports
 import 'src/app/app.dart';
@@ -29,6 +30,8 @@ import 'src/workspace/workspace.dart';
 import 'src/users/users.dart';
 // import 'src/ble/ble.dart';
 import 'src/references/references.dart';
+// import 'src/notifications/notifications.dart';
+import 'src/operations/operations.dart';
 
 // Internal exports
 export 'src/app/app.dart';
@@ -61,6 +64,8 @@ export 'src/workspace/workspace.dart';
 export 'src/users/users.dart';
 export 'src/ble/ble.dart';
 export 'src/references/references.dart';
+export 'src/notifications/notifications.dart';
+export 'src/operations/operations.dart';
 
 // Freezed
 part 'layrz_models.freezed.dart';
@@ -92,7 +97,6 @@ part 'src/guide.dart';
 part 'src/language.dart';
 part 'src/lint_errors.dart';
 part 'src/monitor/checkpoints.dart';
-part 'src/operation.dart';
 part 'src/presets.dart';
 part 'src/report.dart';
 part 'src/suspended_service.dart';