+++ lib/src/operations/operations.dart
@@ -1,0 +1,15 @@
+library operations;
+
+import 'package:collection/collection.dart';
+import 'package:flutter/material.dart';
+import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:layrz_models/layrz_models.dart';
+
+part 'operations.freezed.dart';
+part 'operations.g.dart';
+
+part 'src/operation_type.dart';
+part 'src/http_request_type.dart';
+part 'src/notification_type.dart';
+part 'src/operation.dart';
+part 'src/http_header.dart';